extract to CustomQueryInput()
